Restructure project and add packaging scripts
