+++ src/satellite_remote/cli.py
